Mark up property refs in schema comments.
When processing docs for a Pulumi schema, replace references to resource
properties with JSON Path links to the respective resource or type
properties in the schema. A link to a property looks like this:

```
[`b64Url`](#/resources/random:index%2FrandomId:RandomId/properties/b64Url)
```

This link refers to the `b64Url` output property of the
`random:index/randomId:RandomId` resource. These links can be
interpreted by downstream consumers when generating documentation.